# api/openapi.yaml
openapi: 3.0.3
info:
  title: Etherview API
  version: 0.1.0
  description: Versioned native API for a single-chain Etherview deployment.
servers:
  - url: /api/v1
paths:
  /status:
    get:
      operationId: getStatus
      tags: [system]
      responses:
        "200":
          description: Current indexing and capability status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    APIKey:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    Error:
      description: Machine-readable error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Quantity:
      type: string
      pattern: '^(0|[1-9][0-9]*)$'
      maxLength: 78
      description: A uint256 in the inclusive range 0 through 2^256-1, serialized as a canonical decimal string.
...
    OpaqueCursor:
      type: string
      maxLength: 1024
      description: A bounded server-issued traversal token that clients must not decode or construct.
    Completeness:
      type: object
      required: [core, trace, metadata, state]
      properties:
        core:
          $ref: "#/components/schemas/StageState"
        trace:
          $ref: "#/components/schemas/StageState"
        metadata:
          $ref: "#/components/schemas/StageState"
        state:
          $ref: "#/components/schemas/StageState"
    StageState:
      type: string
      enum: [complete, pending, unavailable, failed]
...
    Meta:
      type: object
      required: [request_id, chain_id]
      properties:
        request_id:
          type: string
        chain_id:
          $ref: "#/components/schemas/Quantity"
        next_cursor:
          $ref: "#/components/schemas/OpaqueCursor"
        coverage_start:
          $ref: "#/components/schemas/Quantity"
        coverage_end:
          $ref: "#/components/schemas/Quantity"
    Status:
      type: object
      required: [chain_id, core_ready, latest_block, indexed_block, backfill_complete, lag, completeness]
      properties:
        chain_id:
          $ref: "#/components/schemas/Quantity"
        core_ready:
          type: boolean
        latest_block:
          $ref: "#/components/schemas/Quantity"
        indexed_block:
          $ref: "#/components/schemas/Quantity"
          description: Highest canonical block contiguous from the configured start block.
        highest_covered_block:
          $ref: "#/components/schemas/Quantity"
          description: Highest canonical block stored in any coverage range; absent before the first block is stored.
        backfill_complete:
          type: boolean
          description: Whether canonical coverage is contiguous from the configured start through the latest known head.
        safe_block:
          $ref: "#/components/schemas/Quantity"
        finalized_block:
          $ref: "#/components/schemas/Quantity"
        lag:
          $ref: "#/components/schemas/Quantity"
        completeness:
          $ref: "#/components/schemas/Completeness"
...
    APIError:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        request_id:
          type: string
    StatusResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Status"
        meta:
          $ref: "#/components/schemas/Meta"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"